requirements.txt
keras==2.11.0
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
streamlit==1.19.0
